Refactor: apollo local only fields to be better separated
