+++ src/cache/colony/index.ts
@@ -1,0 +1,13 @@
+import tokensField from './tokens';
+import balancesField from './balances';
+
+const colony = {
+  Colony: {
+    fields: {
+      ...tokensField,
+      ...balancesField,
+    },
+  },
+};
+
+export default colony;
